Added functionality for the user to enter arbitrary key/num pairs separated with commas. E.g: Enter a key to use in the JSON file: key1,4,key2,3,key3,5
If they want to enter 3 keys that matches a text file with 12 valid lines.

Added the function _resolve_key_list to segregate the logic for parsing user key input.

It will also pad the number if the user entered too few key values or cut it short if there are too many keys.